Convert to function in workflows, no doc
